model hotfix

+++ app/src/main/java/com/dalpak/bringit/models/OrderModel.java
@@ -22,7 +22,7 @@ public class OrderModel {
     @SerializedName("change_type")
     private String mChangeType;
     @SerializedName("total_paid")
-    private int mTotalPaid;
+    private double mTotalPaid;
     @SerializedName("is_paid")
     private int mIsPaid;
     @SerializedName("position")
@@ -70,11 +70,11 @@ public class OrderModel {
         mIsDelivery = isDelivery;
     }
 
-    public int getTotalPaid() {
+    public double getTotalPaid() {
         return mTotalPaid;
     }
 
-    public void setTotalPaid(int totalPaid) {
+    public void setTotalPaid(double totalPaid) {
         mTotalPaid = totalPaid;
     }
